requirements.txt
streamlit
google-generativeai
langchain
pyPDF2
faiss-cpu
langchain_google_genai
python-dotenv